Add temperature mechanic, water areas, temperature gauge
Temperature works with a list of heating/cooling sources and calculates the
heating rate in the process function. At the moment there's
only one heating source: the active weapon.
I tried to make it work with increase/decrease_rate methods but it lead to hard
bugs. The current solution seems a lot more reliable.